Change onboarding permission feature to theme 2
