+++ feature/onboarding/permissions/src/debug/kotlin/app/k9mail/feature/onboarding/permissions/ui/PermissionScreenPreview.kt
@@ -1,0 +1,21 @@
+package app.k9mail.feature.onboarding.permissions.ui
+
+import androidx.compose.runtime.Composable
+import app.k9mail.core.android.permissions.PermissionState
+import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
+import app.k9mail.core.ui.compose.designsystem.PreviewWithTheme
+import kotlinx.coroutines.Dispatchers
+
+@Composable
+@PreviewDevices
+internal fun PermissionScreenPreview() {
+    PreviewWithTheme {
+        PermissionsScreen(
+            viewModel = PermissionsViewModel(
+                checkPermission = { PermissionState.Denied },
+                backgroundDispatcher = Dispatchers.Main.immediate,
+            ),
+            onNext = {},
+        )
+    }
+}
