+++ feature/onboarding/permissions/src/main/kotlin/app/k9mail/feature/onboarding/permissions/ui/PermissionBox.kt
@@ -17,17 +17,14 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import app.k9mail.core.ui.compose.common.image.ImageWithOverlayCoordinate
 import app.k9mail.core.ui.compose.designsystem.atom.button.ButtonFilled
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
-import app.k9mail.core.ui.compose.designsystem.atom.icon.IconsWithBottomRightOverlay
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextBodyMedium
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextTitleLarge
-import app.k9mail.core.ui.compose.theme.K9Theme
-import app.k9mail.core.ui.compose.theme.MainTheme
+import app.k9mail.core.ui.compose.theme2.MainTheme
 import app.k9mail.feature.onboarding.permissions.R
 import app.k9mail.feature.onboarding.permissions.ui.PermissionsContract.UiPermissionState
 
@@ -128,47 +125,5 @@ private fun IconWithPermissionStateOverlay(
                 )
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun PermissionBoxUnknownStatePreview() {
-    K9Theme {
-        PermissionBox(
-            icon = IconsWithBottomRightOverlay.person,
-            permissionState = UiPermissionState.Unknown,
-            title = "Contacts",
-            description = "Allow access to be able to display contact names and photos.",
-            onAllowClick = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun PermissionBoxGrantedStatePreview() {
-    K9Theme {
-        PermissionBox(
-            icon = IconsWithBottomRightOverlay.person,
-            permissionState = UiPermissionState.Granted,
-            title = "Contacts",
-            description = "Allow access to be able to display contact names and photos.",
-            onAllowClick = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun PermissionBoxDeniedStatePreview() {
-    K9Theme {
-        PermissionBox(
-            icon = IconsWithBottomRightOverlay.person,
-            permissionState = UiPermissionState.Denied,
-            title = "Contacts",
-            description = "Allow access to be able to display contact names and photos.",
-            onAllowClick = {},
-        )
     }
 }
