+++ feature/onboarding/permissions/src/main/kotlin/app/k9mail/feature/onboarding/permissions/ui/PermissionsContent.kt
@@ -19,7 +19,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
 import app.k9mail.core.ui.compose.common.visibility.hide
 import app.k9mail.core.ui.compose.designsystem.atom.DelayedCircularProgressIndicator
 import app.k9mail.core.ui.compose.designsystem.atom.Surface
@@ -29,13 +28,11 @@ import app.k9mail.core.ui.compose.designsystem.atom.icon.IconsWithBottomRightOve
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextHeadlineSmall
 import app.k9mail.core.ui.compose.designsystem.template.ResponsiveWidthContainer
 import app.k9mail.core.ui.compose.designsystem.template.Scaffold
-import app.k9mail.core.ui.compose.theme.K9Theme
-import app.k9mail.core.ui.compose.theme.MainTheme
+import app.k9mail.core.ui.compose.theme2.MainTheme
 import app.k9mail.feature.account.common.ui.AppTitleTopHeader
 import app.k9mail.feature.onboarding.permissions.R
 import app.k9mail.feature.onboarding.permissions.ui.PermissionsContract.Event
 import app.k9mail.feature.onboarding.permissions.ui.PermissionsContract.State
-import app.k9mail.feature.onboarding.permissions.ui.PermissionsContract.UiPermissionState
 import app.k9mail.feature.account.common.R as CommonR
 
 @Composable
@@ -179,22 +176,5 @@ private fun BottomBar(
                 }
             }
         }
-    }
-}
-
-@PreviewDevices
-@Composable
-internal fun PermissionContentPreview() {
-    K9Theme {
-        PermissionsContent(
-            state = State(
-                isLoading = false,
-                contactsPermissionState = UiPermissionState.Granted,
-                notificationsPermissionState = UiPermissionState.Denied,
-                isNotificationsPermissionVisible = true,
-                isNextButtonVisible = false,
-            ),
-            onEvent = {},
-        )
     }
 }
